+++ src/Pulse1x.App/Views/GameHub/StatisticsWindow.xaml.cs
@@ -1,4 +1,6 @@
 using Pulse1x.App.ViewModels.GameHub;
+using Pulse1x.App.Services.GameHub;
+using Pulse1x.App.Services;
 using Wpf.Ui.Controls;
 
 namespace Pulse1x.App.Views.GameHub;
@@ -11,5 +13,14 @@ public partial class StatisticsWindow : FluentWindow
         InitializeComponent();
         DataContext = viewModel;
         viewModel.CloseRequested += Close;
+        Loaded += (_, _) =>
+        {
+            Animations.OpenWindow(RootGrid);
+            Dispatcher.BeginInvoke(() =>
+            {
+                if (GamesList.Items.Count > 0) GamepadFocusService.FocusFirst(GamesList);
+                else MasterToggle.Focus();
+            }, System.Windows.Threading.DispatcherPriority.Input);
+        };
     }
 }